Have server build ignore assets
