+++ lib/app/configs/webpack/client.ts
@@ -1,9 +1,10 @@
+import autoprefixer from "autoprefixer";
 import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import OptimizeCSSAssetsPlugin from "optimize-css-assets-webpack-plugin";
 import path from "path";
 import SWPrecacheWebpackPlugin from "sw-precache-webpack-plugin";
 import UglifyJsPlugin from "uglifyjs-webpack-plugin";
-import webpack from "webpack";
+import webpack, {Module, RuleSetRule} from "webpack";
 import {DevtoolModuleFilenameTemplateInfo, Node, Options, Output, Plugin} from "webpack";
 import {getIfUtils, removeEmpty} from "webpack-config-utils";
 import ManifestPlugin from "webpack-manifest-plugin";
@@ -43,6 +44,113 @@ const entry = removeEmpty([
     ifDevelopment("webpack-hot-middleware/client?reload=true", undefined),
     resolvePath("src/client/index.tsx"),
 ]);
+
+/**
+ * Post CSS fixes
+ */
+const postCssPlugin = () => [
+    require("postcss-flexbugs-fixes"),
+    autoprefixer({
+        browsers: [
+            ">1%",
+            "last 4 versions",
+            "Firefox ESR",
+            "not ie < 9", // React doesn"t support IE8 anyway
+        ],
+        flexbox: "no-2009",
+    }),
+];
+
+/**
+ * All other files that aren't caught by the other loaders will go through this one.
+ * @description "file" loader makes sure those assets get served by WebpackDevServer.
+ * When you `import` an asset, you get its (virtual) filename.
+ * In production, they would get copied to the `build` folder.
+ * This loader doesn"t use a "test" so it will catch all modules
+ * that fall through the other loaders.
+ */
+const catchAllRule = {
+    exclude: [/\.(js|jsx|mjs)$/, /\.(ts|tsx)$/, /\.html$/, /\.json$/],
+    loader: require.resolve("file-loader"),
+    options: {
+        name: "[name].[hash:8].[ext]",
+        outputPath: "static/media/", // TODO config
+    },
+};
+
+/**
+ * Handle css/scss
+ */
+// TODO do we need source maps here?
+const scssRule: RuleSetRule = {
+    test: /\.(scss|sass|css)$/i,
+    use: removeEmpty([
+        ifProduction(MiniCssExtractPlugin.loader, "style-loader"),
+        {loader: "css-loader"},
+        {
+            loader: "postcss-loader",
+            options: {
+                ident: "postcss",
+                plugins: postCssPlugin,
+            },
+        },
+        {loader: "sass-loader"},
+    ]),
+};
+
+/**
+ * Define rule for static assets
+ * @description "url" loader works like "file" loader except that it embeds assets
+ * smaller than specified limit in bytes as data URLs to avoid requests.
+ */
+const staticFileRule: RuleSetRule = {
+    loader: require.resolve("url-loader"),
+    options: {
+        limit: 10000,
+        name: "static/media/[name].[hash:8].[ext]", // TODO is this needed I don't think so
+    },
+    test: /\.(bmp|png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
+};
+
+/**
+ * Define rule for transpiling TypeScript
+ * @description Uncomment transpileOnly to Disable type checker - will use it in ForkTsCheckerWebpackPlugin at the cost of overlay.
+ * Babel loader is present to support react-hot-loader.
+ *
+ * @todo Make configurable for CI and performance. Babel can also provide caching and polyfill
+ * @todo Babel probably doesn't need to be run for server config
+ */
+const typeScriptRule: RuleSetRule = {
+    include: resolvePath("src"),
+    test: /\.(ts|tsx)$/,
+    use: [
+        {
+            loader: "babel-loader",
+            options: {
+                babelrc: false,
+                cacheDirectory: true,
+                plugins: ["react-hot-loader/babel"],
+            },
+        },
+        {
+            loader: "ts-loader",
+            options: {
+                configFile: resolvePath("tsconfig.json"),
+                // transpileOnly: true,
+            },
+        },
+    ],
+};
+
+/**
+ * Define how source files are handled
+ */
+const module: Module = {
+    rules: [{
+        oneOf: [staticFileRule, typeScriptRule, scssRule, catchAllRule],
+    }],
+    strictExportPresence: true,
+};
 
 /**
  * @description Some libraries import Node modules but don"t use them in the browser.
@@ -168,6 +276,7 @@ const plugins: Plugin[] = removeEmpty([
 export default merge(common, {
     devtool: ifProduction("source-map", "eval-source-map"), // TODO if prod should we even do this???
     entry,
+    module,
     name: "client",
     node,
     optimization,
